feat(agent): convert AI settings into AI & Agents master-detail shell
